Changed to new NC app skeleton structure
