+++ appinfo/app.php
@@ -1,34 +1,25 @@
 <?php
 /**
- * ownCloud - polls
- *
- * This file is licensed under the Affero General Public License version 3 or
- * later. See the COPYING file.
+ * @copyright Copyright (c) 2017 Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
  *
  * @author Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
- * @copyright Vinzenz Rosenkranz 2016
+ *
+ * @license GNU AGPL version 3 or any later version
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
-namespace OCA\Polls\AppInfo;
-
-$l = \OC::$server->getL10N('polls');
-
-\OC::$server->getNavigationManager()->add(array(
-    // the string under which your app will be referenced in owncloud
-    'id' => 'polls',
-
-    // sorting weight for the navigation. The higher the number, the higher
-    // will it be listed in the navigation
-    'order' => 77,
-
-    // the route that will be shown on startup
-    'href' => \OC::$server->getURLGenerator()->linkToRoute('polls.page.index'),
-
-    // the icon that will be shown in the navigation
-    // this file needs to exist in img/
-    'icon' => \OC::$server->getURLGenerator()->imagePath('polls', 'app-logo-polls.svg'),
-
-    // the title of your application. This will be used in the
-    // navigation or on the settings page of your app
-    'name' => $l->t('Polls')
-));
+$app = new \OCA\Polls\AppInfo\Application();
+$app->registerNavigationEntry();
