+++ appinfo/routes.php
@@ -1,25 +1,28 @@
 <?php
 /**
- * This file is licensed under the Affero General Public License version 3 or later.
- * See the COPYING-README file.
+ * @copyright Copyright (c) 2017 Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
  *
  * @author Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
- * @copyright Vinzenz Rosenkranz 2016
- */
-
-namespace OCA\Polls\AppInfo;
-
-/**
- * Create your routes in here. The name is the lowercase name of the controller
- * without the controller part, the stuff after the hash is the method.
- * e.g. page#index -> PageController->index()
  *
- * The controller class has to be registered in the application.php file since
- * it's instantiated in there
+ * @license GNU AGPL version 3 or any later version
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
-$application = new Application();
 
-$application->registerRoutes($this, array('routes' => array(
+$app = new \OCA\Polls\AppInfo\Application();
+$app->registerRoutes($this, array('routes' => array(
 	array('name' => 'page#index', 'url' => '/', 'verb' => 'GET'),
 	array('name' => 'page#goto_poll', 'url' => '/poll/{hash}', 'verb' => 'GET'),
 	array('name' => 'page#edit_poll', 'url' => '/edit/{hash}', 'verb' => 'GET'),
