+++ lib/Db/Access.php
@@ -21,11 +21,17 @@
  *
  */
 
-\OC::$server->getNavigationManager()->setActiveEntry( 'polls' );
+namespace OCA\Polls\Db;
 
-if (OCP\User::isLoggedIn()) {
-    $tmpl = new OCP\Template('polls', 'main', 'user');
-} else {
-    $tmpl = new OCP\Template('polls', 'main', 'base');
+use OCP\AppFramework\Db\Entity;
+
+/**
+ * @method integer getPId()
+ * @method void setPId(integer $value)
+ * @method string getAccessType()
+ * @method void setAccessType(string $value)
+ */
+class Access extends Entity {
+    public $pId;
+    public $accessType;
 }
-$tmpl->printPage();
