+++ lib/Db/Comment.php
@@ -21,11 +21,23 @@
  *
  */
 
-\OC::$server->getNavigationManager()->setActiveEntry( 'polls' );
+namespace OCA\Polls\Db;
 
-if (OCP\User::isLoggedIn()) {
-    $tmpl = new OCP\Template('polls', 'main', 'user');
-} else {
-    $tmpl = new OCP\Template('polls', 'main', 'base');
+use OCP\AppFramework\Db\Entity;
+
+/**
+ * @method string getUserId()
+ * @method void setUserId(string $value)
+ * @method string getDt()
+ * @method void setDt(string $value)
+ * @method string getComment()
+ * @method void setComment(string $value)
+ * @method integer getPollId()
+ * @method void setPollId(integer $value)
+ */
+class Comment extends Entity {
+    public $userId;
+    public $dt;
+    public $comment;
+    public $pollId;
 }
-$tmpl->printPage();
