+++ lib/Db/Date.php
@@ -21,11 +21,17 @@
  *
  */
 
-\OC::$server->getNavigationManager()->setActiveEntry( 'polls' );
+namespace OCA\Polls\Db;
 
-if (OCP\User::isLoggedIn()) {
-    $tmpl = new OCP\Template('polls', 'main', 'user');
-} else {
-    $tmpl = new OCP\Template('polls', 'main', 'base');
+use OCP\AppFramework\Db\Entity;
+
+/**
+ * @method timestamp getDt()
+ * @method void setDt(timestamp $value)
+ * @method integer getPollId()
+ * @method void setPollId(integer $value)
+ */
+class Date extends Entity {
+    public $dt;
+    public $pollId;
 }
-$tmpl->printPage();
