+++ lib/Db/Notification.php
@@ -21,11 +21,17 @@
  *
  */
 
-\OC::$server->getNavigationManager()->setActiveEntry( 'polls' );
+namespace OCA\Polls\Db;
 
-if (OCP\User::isLoggedIn()) {
-    $tmpl = new OCP\Template('polls', 'main', 'user');
-} else {
-    $tmpl = new OCP\Template('polls', 'main', 'base');
+use OCP\AppFramework\Db\Entity;
+
+/**
+ * @method string getUserId()
+ * @method void setUserId(string $value)
+ * @method string getPollId()
+ * @method void setPollId(string $value)
+ */
+class Notification extends Entity {
+    public $userId;
+    public $pollId;
 }
-$tmpl->printPage();
