+++ templates/goto.tmpl.php
@@ -1,4 +1,25 @@
 <?php
+    /**
+     * @copyright Copyright (c) 2017 Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
+     *
+     * @author Vinzenz Rosenkranz <vinzenz.rosenkranz@gmail.com>
+     *
+     * @license GNU AGPL version 3 or any later version
+     *
+     *  This program is free software: you can redistribute it and/or modify
+     *  it under the terms of the GNU Affero General Public License as
+     *  published by the Free Software Foundation, either version 3 of the
+     *  License, or (at your option) any later version.
+     *
+     *  This program is distributed in the hope that it will be useful,
+     *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+     *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     *  GNU Affero General Public License for more details.
+     *
+     *  You should have received a copy of the GNU Affero General Public License
+     *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+     *
+     */
 
     use \OCP\User;
 
@@ -58,7 +79,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
         <div id="app-content-wrapper" class="<?php p($layout); ?>">
 			<div id="controls">
                 <div id="breadcrump">
-                    <?php if(User::isLoggedIn()) : ?>
+                    <?php if (User::isLoggedIn()) : ?>
                     <div class="crumb svg" data-dir="/">
                         <a href="<?php p($urlGenerator->linkToRoute('polls.page.index')); ?>">
                             <img class="svg" src="<?php print_unescaped(OCP\image_path("core", "places/home.svg")); ?>"" alt="Home">
@@ -69,7 +90,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                     </div>
                     <?php endif; ?>
 		
-                    <?php if(!User::isLoggedIn()) : ?>
+                    <?php if (!User::isLoggedIn()) : ?>
                     <div class="col-100">
                         <h2><?php p($poll->getTitle()); ?></h2>
                     </div>
@@ -120,13 +141,13 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                                         }
                                         print_unescaped('<tr>');
                                         print_unescaped('<td class="avatar-cell">');
-                                        if($userMgr->get($usr) != null && !$isAnonymous && !$hideNames) {
+                                        if ($userMgr->get($usr) != null && !$isAnonymous && !$hideNames) {
                                             print_unescaped('<div class="poll avatardiv" title="'.($usr).'"></div>');
                                             print_unescaped('</td>');
                                             print_unescaped('<td colspan="2" class="name">');
                                             p($userMgr->get($usr)->getDisplayName());
                                         } else {
-                                            if($isAnonymous || $hideNames) {
+                                            if ($isAnonymous || $hideNames) {
                                             print_unescaped('<div class="poll avatardiv" title="'.($userCnt).'"></div>');
                                             print_unescaped('</td>');
                                             print_unescaped('<td colspan="2" class="name">');
@@ -141,7 +162,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
 
                                         // loop over dts
                                         $i_tot = 0;
-                                        foreach($dates as $dt) {
+                                        foreach ($dates as $dt) {
                                             if ($poll->getType() == '0') {
                                                 $date_id = strtotime($dt->getDt());
                                                 $poll_id = "pollid_" . $dt->getId();
@@ -153,7 +174,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                                             $found = false;
                                             foreach ($others[$usr] as $vote) {
                                                 $voteVal = null;
-                                                if($poll->getType() == '0') {
+                                                if ($poll->getType() == '0') {
                                                     $voteVal = strtotime($vote->getDt());
                                                 } else {
                                                     $voteVal = $vote->getText();
@@ -174,7 +195,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                                                     break;
                                                 }
                                             }
-                                            if(!$found) {
+                                            if (!$found) {
                                                 $cl = 'poll-cell unvoted';
                                             }
                                             // Make the td clickable
@@ -298,9 +319,9 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                             <input type="hidden" name="changed" />
                             <input type="button" id="submit_finish_vote" class="button btn" value="<?php p($l->t('Vote!')); ?>" />
                         </form>
-                        <?php if(User::isLoggedIn()) : ?>
+                        <?php if (User::isLoggedIn()) : ?>
                         <span class="notification">
-                            <input type="checkbox" id="check_notif" class="checkbox" <?php if($notification !== null) print_unescaped(' checked'); ?> />
+                            <input type="checkbox" id="check_notif" class="checkbox" <?php if ($notification !== null) print_unescaped(' checked'); ?> />
                             <label for="check_notif"><?php p($l->t('Receive notification email on activity')); ?></label>
                         </span>
                         <?php endif; ?>
@@ -316,7 +337,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
 
 
 
-                    <?php if($expired) : ?>
+                    <?php if ($expired) : ?>
                         <div id="expired_info">
                             <h2><?php p($l->t('Poll expired')); ?></h2>
                             <p>
@@ -331,7 +352,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                                 <input type="hidden" name="pollId" value="<?php p($poll->getId()); ?>" />
                                 <input type="hidden" name="userId" value="<?php p($userId); ?>" />
                                 <div class="comment-content">
-                                <?php if(!User::isLoggedIn()) : ?>
+                                <?php if (!User::isLoggedIn()) : ?>
                                     <!--<?php
                                     p($l->t('You must be logged in to post a comment.'));
                                     ?>-->
@@ -349,16 +370,16 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
                                 </div>
                             </form>
                         </div>
-                        <?php if($comments !== null) : ?>
+                        <?php if ($comments !== null) : ?>
                             <?php foreach ($comments as $comment) : ?>
                                 <div class="comment">
                                     <div class="comment-header">
                                         <?php
                                         print_unescaped('<span class="comment-date">' . date('d.m.Y H:i:s', strtotime($comment->getDt())) . '</span>');
-                                        if($isAnonymous || $hideNames) {
+                                        if ($isAnonymous || $hideNames) {
                                             p('Anonymous');
                                         } else {
-                                            if($userMgr->get($comment->getUserId()) != null) {
+                                            if ($userMgr->get($comment->getUserId()) != null) {
                                                 p($userMgr->get($comment->getUserId())->getDisplayName());
                                             } else {
                                                 print_unescaped('<i>');
@@ -388,7 +409,7 @@ $pollUrl = $urlGenerator->linkToRouteAbsolute('polls.page.goto_poll', ['hash' =>
 //adapted from jsxc.chat
 function getHsl($str) {
     $hash = 0;
-    for($i=0; $i<strlen($str); $i++) {
+    for ($i=0; $i<strlen($str); $i++) {
         $utf16_char = mb_convert_encoding($str[$i], "utf-16", "utf-8");
         $char = hexdec(bin2hex($utf16_char));
         $hash = (($hash << 5) - $hash) + $char;
